+++ java/src/com/ctre/phoenix/Drive/ISmartDrivetrain.java
@@ -2,12 +2,12 @@ package com.ctre.phoenix.Drive;
 
 public interface ISmartDrivetrain extends IDrivetrain
 {
-	void set(Styles.Smart mode, float forward, float turn);
+	void set(Styles.Smart mode, double forward, double turn);
 
     void SetCurrentLimit(int currentLimitAmps, int timeoutMs);
-    float GetDistance();
-    float GetVelocity();
+    double GetDistance();
+    double GetVelocity();
 
-    float GetEncoderHeading();
-    void SetPosition(float position);
+    double GetEncoderHeading();
+    void SetPosition(double position);
 }
